refactor(Charts): Extract common logic into reusable hooks
