+++ packages/charts/src/components/DonutChart/index.tsx
@@ -1,9 +1,10 @@
 import { useConsolidatedRef } from '@ui5/webcomponents-react-base';
-import React, { forwardRef, Ref, RefObject, useCallback, useEffect, useRef, useMemo } from 'react';
+import React, { forwardRef, Ref, RefObject, useMemo, useRef } from 'react';
 import { Pie } from 'react-chartjs-2';
 import { useTheme } from 'react-jss';
 import { ChartBaseProps } from '../../interfaces/ChartBaseProps';
-import { withChartContainer } from '../../internal/ChartContainer/withChartContainer';
+import { useLegend, usePieLegendItemClickHandler } from '../../internal/ChartLegend';
+import { withChartContainer } from '../../internal/withChartContainer';
 import { ChartBaseDefaultProps } from '../../util/ChartBaseDefaultProps';
 import { useChartData } from '../../util/populateData';
 import { formatTooltipLabelForPieCharts, useMergedConfig } from '../../util/utils';
@@ -55,29 +56,8 @@ const DonutChartComponent = forwardRef((props: DonutChartPropTypes, ref: Ref<any
   const chartRef = useConsolidatedRef<any>(ref);
   const legendRef: RefObject<HTMLDivElement> = useRef();
 
-  const handleLegendItemPress = useCallback(
-    (e) => {
-      const clickTarget = (e.currentTarget as unknown) as HTMLLIElement;
-      const datasetIndex = parseInt(clickTarget.dataset.datasetindex);
-      const { chartInstance } = chartRef.current;
-      const meta = chartInstance.getDatasetMeta(0).data[datasetIndex];
-      meta.hidden = !meta.hidden;
-      chartInstance.update();
-      clickTarget.style.textDecoration = meta.hidden ? 'line-through' : 'unset';
-    },
-    [legendRef.current, chartRef.current]
-  );
-
-  useEffect(() => {
-    if (noLegend) {
-      legendRef.current.innerHTML = '';
-    } else {
-      legendRef.current.innerHTML = chartRef.current.chartInstance.generateLegend();
-      legendRef.current.querySelectorAll('li').forEach((legendItem) => {
-        legendItem.addEventListener('click', handleLegendItemPress);
-      });
-    }
-  }, [chartRef.current, legendRef.current, noLegend]);
+  const handleLegendItemPress = usePieLegendItemClickHandler(chartRef, legendRef);
+  useLegend(chartRef, legendRef, noLegend, handleLegendItemPress);
 
   return (
     <>
