+++ packages/charts/src/components/LineChart/index.tsx
@@ -1,10 +1,11 @@
 import { useConsolidatedRef } from '@ui5/webcomponents-react-base';
-import React, { forwardRef, Ref, RefObject, useCallback, useEffect, useRef, useMemo } from 'react';
+import React, { forwardRef, Ref, RefObject, useMemo, useRef } from 'react';
 import { Line } from 'react-chartjs-2';
 import { useTheme } from 'react-jss';
 import { DEFAULT_OPTIONS } from '../../config';
 import { ChartBaseProps } from '../../interfaces/ChartBaseProps';
-import { withChartContainer } from '../../internal/ChartContainer/withChartContainer';
+import { useLegend, useLegendItemClickHandler } from '../../internal/ChartLegend';
+import { withChartContainer } from '../../internal/withChartContainer';
 import { ChartBaseDefaultProps } from '../../util/ChartBaseDefaultProps';
 import { useChartData } from '../../util/populateData';
 import { formatTooltipLabel, useMergedConfig } from '../../util/utils';
@@ -61,29 +62,8 @@ const LineChartComponent = forwardRef((props: LineChartPropTypes, ref: Ref<any>)
 
   const chartRef = useConsolidatedRef<any>(ref);
   const legendRef: RefObject<HTMLDivElement> = useRef();
-  const handleLegendItemPress = useCallback(
-    (e) => {
-      const clickTarget = (e.currentTarget as unknown) as HTMLLIElement;
-      const datasetIndex = parseInt(clickTarget.dataset.datasetindex);
-      const { chartInstance } = chartRef.current;
-      const meta = chartInstance.getDatasetMeta(datasetIndex);
-      meta.hidden = meta.hidden === null ? !chartInstance.data.datasets[datasetIndex].hidden : null;
-      chartInstance.update();
-      clickTarget.style.textDecoration = meta.hidden ? 'line-through' : 'unset';
-    },
-    [legendRef.current, chartRef.current]
-  );
-
-  useEffect(() => {
-    if (noLegend) {
-      legendRef.current.innerHTML = '';
-    } else {
-      legendRef.current.innerHTML = chartRef.current.chartInstance.generateLegend();
-      legendRef.current.querySelectorAll('li').forEach((legendItem) => {
-        legendItem.addEventListener('click', handleLegendItemPress);
-      });
-    }
-  }, [chartRef.current, legendRef.current, noLegend]);
+  const handleLegendItemPress = useLegendItemClickHandler(chartRef, legendRef);
+  useLegend(chartRef, legendRef, noLegend, handleLegendItemPress);
 
   return (
     <>
