+++ packages/charts/src/internal/ChartLegend.ts
@@ -1,4 +1,5 @@
 import { Optional } from '@ui5/webcomponents-react-base';
+import { useEffect, Ref, RefObject, useCallback } from 'react';
 
 const getSymbolForDataset = (type, dataset) => {
   const datasetMeta: any = dataset.hasOwnProperty('_meta') ? Object.values(dataset._meta)[0] : null;
@@ -36,4 +37,54 @@ export const generateLegend = (chart) => {
   return `<div style="display: flex; flex-wrap: wrap; padding: 0 1rem; box-sizing: border-box">
 ${itemsForLegend.map((item, index) => getLegendEntry(item, type, index)).join(' ')}
 </div>`;
+};
+
+export const useLegend = (
+  chartRef: RefObject<any>,
+  legendRef: RefObject<HTMLElement>,
+  noLegend: boolean,
+  handleLegendItemPress: (e: any) => void
+) => {
+  useEffect(() => {
+    if (chartRef && legendRef) {
+      if (noLegend) {
+        legendRef.current.innerHTML = '';
+      } else {
+        legendRef.current.innerHTML = chartRef.current.chartInstance.generateLegend();
+        legendRef.current.querySelectorAll('li').forEach((legendItem) => {
+          legendItem.addEventListener('click', handleLegendItemPress);
+        });
+      }
+    }
+  }, [chartRef.current, legendRef.current, noLegend, handleLegendItemPress]);
+};
+
+export const usePieLegendItemClickHandler = (chartRef: RefObject<any>, legendRef: RefObject<HTMLElement>) => {
+  return useCallback(
+    (e) => {
+      const clickTarget = (e.currentTarget as unknown) as HTMLLIElement;
+      const datasetIndex = parseInt(clickTarget.dataset.datasetindex);
+      const { chartInstance } = chartRef.current;
+      const meta = chartInstance.getDatasetMeta(0).data[datasetIndex];
+      meta.hidden = !meta.hidden;
+      chartInstance.update();
+      clickTarget.style.textDecoration = meta.hidden ? 'line-through' : 'unset';
+    },
+    [legendRef.current, chartRef.current]
+  );
+};
+
+export const useLegendItemClickHandler = (chartRef: RefObject<any>, legendRef: RefObject<HTMLElement>) => {
+  return useCallback(
+    (e) => {
+      const clickTarget = (e.currentTarget as unknown) as HTMLLIElement;
+      const datasetIndex = parseInt(clickTarget.dataset.datasetindex);
+      const { chartInstance } = chartRef.current;
+      const meta = chartInstance.getDatasetMeta(datasetIndex);
+      meta.hidden = meta.hidden === null ? !chartInstance.data.datasets[datasetIndex].hidden : null;
+      chartInstance.update();
+      clickTarget.style.textDecoration = meta.hidden ? 'line-through' : 'unset';
+    },
+    [legendRef.current, chartRef.current]
+  );
 };
