+++ packages/charts/src/themes/themeMap.ts
@@ -9,15 +9,6 @@ themeMap.set('sap_belize_plus', sap_belize_plus);
 themeMap.set('sap_fiori_3', sap_fiori_3);
 themeMap.set('sap_fiori_3_dark', sap_fiori_3_dark);
 
-// export const getOrLoadTheme = async (theme) => {
-//   if (themeMap.has(theme)) {
-//     return themeMap.get(theme);
-//   }
-//   const { semanticColors, sequentialColors } = await import(`./${theme}`);
-//   themeMap.set(theme, { semanticColors, sequentialColors });
-//   return { semanticColors, sequentialColors };
-// };
-
 export const getOrLoadTheme = (theme) => {
   return themeMap.get(theme);
 };
